cloud_storage: move util functions for partition manifest into partition_manifest file
