--- src/v/cloud_storage/base_manifest.cc
@@ -1,137 +1,0 @@
-/*
- * Copyright 2021 Vectorized, Inc.
- *
- * Licensed as a Redpanda Enterprise file under the Redpanda Community
- * License (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * https://github.com/vectorizedio/redpanda/blob/master/licenses/rcl.md
- */
-
-#include "cloud_storage/base_manifest.h"
-
-#include "hashing/xx.h"
-#include "ssx/sformat.h"
-#include "storage/fs_utils.h"
-
-#include <seastar/core/coroutine.hh>
-
-#include <fmt/ostream.h>
-
-#include <charconv>
-
-namespace cloud_storage {
-
-std::ostream&
-operator<<(std::ostream& s, const partition_manifest_path_components& c) {
-    fmt::print(
-      s, "{{{}: {}-{}-{}-{}}}", c._origin, c._ns, c._topic, c._part, c._rev);
-    return s;
-}
-
-static bool parse_partition_and_revision(
-  std::string_view s, partition_manifest_path_components& comp) {
-    auto pos = s.find('_');
-    if (pos == std::string_view::npos) {
-        // Invalid segment file name
-        return false;
-    }
-    uint64_t res = 0;
-    // parse first component
-    auto sv = s.substr(0, pos);
-    auto e = std::from_chars(sv.data(), sv.data() + sv.size(), res);
-    if (e.ec != std::errc()) {
-        return false;
-    }
-    comp._part = model::partition_id(res);
-    // parse second component
-    sv = s.substr(pos + 1);
-    e = std::from_chars(sv.data(), sv.data() + sv.size(), res);
-    if (e.ec != std::errc()) {
-        return false;
-    }
-    comp._rev = model::initial_revision_id(res);
-    return true;
-}
-
-std::optional<partition_manifest_path_components>
-get_partition_manifest_path_components(const std::filesystem::path& path) {
-    // example: b0000000/meta/kafka/redpanda-test/4_2/manifest.json
-    enum {
-        ix_prefix,
-        ix_meta,
-        ix_namespace,
-        ix_topic,
-        ix_part_rev,
-        ix_file_name,
-        total_components
-    };
-    partition_manifest_path_components res;
-    res._origin = path;
-    int ix = 0;
-    for (const auto& c : path) {
-        ss::sstring p = c.string();
-        switch (ix++) {
-        case ix_prefix:
-            break;
-        case ix_namespace:
-            res._ns = model::ns(std::move(p));
-            break;
-        case ix_topic:
-            res._topic = model::topic(std::move(p));
-            break;
-        case ix_part_rev:
-            if (!parse_partition_and_revision(p, res)) {
-                return std::nullopt;
-            }
-            break;
-        case ix_file_name:
-            if (p != "manifest.json") {
-                return std::nullopt;
-            }
-            break;
-        }
-    }
-    if (ix == total_components) {
-        return res;
-    }
-    return std::nullopt;
-}
-
-std::optional<segment_name_components>
-parse_segment_name(const segment_name& name) {
-    auto parsed = storage::segment_path::parse_segment_filename(name);
-    if (!parsed) {
-        return std::nullopt;
-    }
-    return segment_name_components{
-      .base_offset = parsed->base_offset,
-      .term = parsed->term,
-    };
-}
-
-remote_segment_path generate_remote_segment_path(
-  const model::ntp& ntp,
-  model::initial_revision_id rev_id,
-  const segment_name& name,
-  model::term_id archiver_term) {
-    vassert(
-      rev_id != model::initial_revision_id(),
-      "ntp {}: ntp revision must be known for segment {}",
-      ntp,
-      name);
-
-    auto path = ssx::sformat("{}_{}/{}", ntp.path(), rev_id(), name());
-    uint32_t hash = xxhash_32(path.data(), path.size());
-    if (archiver_term != model::term_id{}) {
-        return remote_segment_path(
-          fmt::format("{:08x}/{}.{}", hash, path, archiver_term()));
-    } else {
-        return remote_segment_path(fmt::format("{:08x}/{}", hash, path));
-    }
-}
-
-segment_name generate_segment_name(model::offset o, model::term_id t) {
-    return segment_name(ssx::sformat("{}-{}-v1.log", o(), t()));
-}
-} // namespace cloud_storage
