--- src/v/cloud_storage/tests/base_manifest_test.cc
@@ -1,95 +1,0 @@
-/*
- * Copyright 2021 Vectorized, Inc.
- *
- * Licensed as a Redpanda Enterprise file under the Redpanda Community
- * License (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * https://github.com/vectorizedio/redpanda/blob/master/licenses/rcl.md
- */
-
-#include "cloud_storage/base_manifest.h"
-#include "cloud_storage/types.h"
-#include "model/metadata.h"
-#include "seastarx.h"
-
-#include <seastar/testing/test_case.hh>
-#include <seastar/testing/thread_test_case.hh>
-
-#include <boost/test/tools/old/interface.hpp>
-#include <boost/test/unit_test.hpp>
-
-using namespace cloud_storage;
-
-static const model::ntp manifest_ntp(
-  model::ns("test-ns"), model::topic("test-topic"), model::partition_id(42));
-
-SEASTAR_THREAD_TEST_CASE(test_segment_path) {
-    auto path = generate_remote_segment_path(
-      manifest_ntp,
-      model::initial_revision_id(0),
-      segment_name("22-11-v1.log"),
-      model::term_id{123});
-    // use pre-calculated murmur hash value from full ntp path + file name
-    BOOST_REQUIRE_EQUAL(
-      path, "2bea9275/test-ns/test-topic/42_0/22-11-v1.log.123");
-}
-
-SEASTAR_THREAD_TEST_CASE(test_manifest_name_parsing) {
-    std::filesystem::path path
-      = "b0000000/meta/kafka/redpanda-test/4_2/manifest.json";
-    auto res = cloud_storage::get_partition_manifest_path_components(path);
-    BOOST_REQUIRE_EQUAL(res->_origin, path);
-    BOOST_REQUIRE_EQUAL(res->_ns(), "kafka");
-    BOOST_REQUIRE_EQUAL(res->_topic(), "redpanda-test");
-    BOOST_REQUIRE_EQUAL(res->_part(), 4);
-    BOOST_REQUIRE_EQUAL(res->_rev(), 2);
-}
-
-SEASTAR_THREAD_TEST_CASE(test_manifest_name_parsing_failure_1) {
-    std::filesystem::path path
-      = "b0000000/meta/kafka/redpanda-test/a_b/manifest.json";
-    auto res = cloud_storage::get_partition_manifest_path_components(path);
-    BOOST_REQUIRE(res.has_value() == false);
-}
-
-SEASTAR_THREAD_TEST_CASE(test_manifest_name_parsing_failure_2) {
-    std::filesystem::path path
-      = "b0000000/kafka/redpanda-test/4_2/manifest.json";
-    auto res = cloud_storage::get_partition_manifest_path_components(path);
-    BOOST_REQUIRE(res.has_value() == false);
-}
-
-SEASTAR_THREAD_TEST_CASE(test_manifest_name_parsing_failure_3) {
-    std::filesystem::path path
-      = "b0000000/meta/kafka/redpanda-test//manifest.json";
-    auto res = cloud_storage::get_partition_manifest_path_components(path);
-    BOOST_REQUIRE(res.has_value() == false);
-}
-
-SEASTAR_THREAD_TEST_CASE(test_manifest_name_parsing_failure_4) {
-    std::filesystem::path path
-      = "b0000000/meta/kafka/redpanda-test/4_2/foo.bar";
-    auto res = cloud_storage::get_partition_manifest_path_components(path);
-    BOOST_REQUIRE(res.has_value() == false);
-}
-
-SEASTAR_THREAD_TEST_CASE(test_segment_name_parsing) {
-    segment_name name{"3587-1-v1.log"};
-    auto res = parse_segment_name(name);
-    BOOST_REQUIRE(res);
-    BOOST_REQUIRE_EQUAL(res->base_offset(), 3587);
-    BOOST_REQUIRE_EQUAL(res->term(), 1);
-}
-
-SEASTAR_THREAD_TEST_CASE(test_segment_name_parsing_failure_1) {
-    segment_name name{"-1-v1.log"};
-    auto res = parse_segment_name(name);
-    BOOST_REQUIRE(res.has_value() == false);
-}
-
-SEASTAR_THREAD_TEST_CASE(test_segment_name_parsing_failure_2) {
-    segment_name name{"abc-1-v1.log"};
-    auto res = parse_segment_name(name);
-    BOOST_REQUIRE(res.has_value() == false);
-}
