+++ src/v/cloud_storage/tests/partition_manifest_test.cc
@@ -70,6 +70,76 @@ inline ss::input_stream<char> make_manifest_stream(std::string_view json) {
     return make_iobuf_input_stream(std::move(i));
 }
 
+SEASTAR_THREAD_TEST_CASE(test_segment_path) {
+    auto path = generate_remote_segment_path(
+      manifest_ntp,
+      model::initial_revision_id(0),
+      segment_name("22-11-v1.log"),
+      model::term_id{123});
+    // use pre-calculated murmur hash value from full ntp path + file name
+    BOOST_REQUIRE_EQUAL(
+      path, "2bea9275/test-ns/test-topic/42_0/22-11-v1.log.123");
+}
+
+SEASTAR_THREAD_TEST_CASE(test_manifest_name_parsing) {
+    std::filesystem::path path
+      = "b0000000/meta/kafka/redpanda-test/4_2/manifest.json";
+    auto res = cloud_storage::get_partition_manifest_path_components(path);
+    BOOST_REQUIRE_EQUAL(res->_origin, path);
+    BOOST_REQUIRE_EQUAL(res->_ns(), "kafka");
+    BOOST_REQUIRE_EQUAL(res->_topic(), "redpanda-test");
+    BOOST_REQUIRE_EQUAL(res->_part(), 4);
+    BOOST_REQUIRE_EQUAL(res->_rev(), 2);
+}
+
+SEASTAR_THREAD_TEST_CASE(test_manifest_name_parsing_failure_1) {
+    std::filesystem::path path
+      = "b0000000/meta/kafka/redpanda-test/a_b/manifest.json";
+    auto res = cloud_storage::get_partition_manifest_path_components(path);
+    BOOST_REQUIRE(res.has_value() == false);
+}
+
+SEASTAR_THREAD_TEST_CASE(test_manifest_name_parsing_failure_2) {
+    std::filesystem::path path
+      = "b0000000/kafka/redpanda-test/4_2/manifest.json";
+    auto res = cloud_storage::get_partition_manifest_path_components(path);
+    BOOST_REQUIRE(res.has_value() == false);
+}
+
+SEASTAR_THREAD_TEST_CASE(test_manifest_name_parsing_failure_3) {
+    std::filesystem::path path
+      = "b0000000/meta/kafka/redpanda-test//manifest.json";
+    auto res = cloud_storage::get_partition_manifest_path_components(path);
+    BOOST_REQUIRE(res.has_value() == false);
+}
+
+SEASTAR_THREAD_TEST_CASE(test_manifest_name_parsing_failure_4) {
+    std::filesystem::path path
+      = "b0000000/meta/kafka/redpanda-test/4_2/foo.bar";
+    auto res = cloud_storage::get_partition_manifest_path_components(path);
+    BOOST_REQUIRE(res.has_value() == false);
+}
+
+SEASTAR_THREAD_TEST_CASE(test_segment_name_parsing) {
+    segment_name name{"3587-1-v1.log"};
+    auto res = parse_segment_name(name);
+    BOOST_REQUIRE(res);
+    BOOST_REQUIRE_EQUAL(res->base_offset(), 3587);
+    BOOST_REQUIRE_EQUAL(res->term(), 1);
+}
+
+SEASTAR_THREAD_TEST_CASE(test_segment_name_parsing_failure_1) {
+    segment_name name{"-1-v1.log"};
+    auto res = parse_segment_name(name);
+    BOOST_REQUIRE(res.has_value() == false);
+}
+
+SEASTAR_THREAD_TEST_CASE(test_segment_name_parsing_failure_2) {
+    segment_name name{"abc-1-v1.log"};
+    auto res = parse_segment_name(name);
+    BOOST_REQUIRE(res.has_value() == false);
+}
+
 SEASTAR_THREAD_TEST_CASE(test_manifest_path) {
     partition_manifest m(manifest_ntp, model::initial_revision_id(0));
     auto path = m.get_manifest_path();
